WIP: refactoring backend
